requirements.txt
amqp==2.3.2
apt-xapian-index==0.45
billiard==3.5.0.4
boto==2.20.1
celery==4.2.1
chardet==2.0.1
Cheetah==2.4.4
click==6.7
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
crcmod==1.7
Django==1.11.15
djangorestframework==3.8.2
google-compute-engine==2.8.2
gunicorn==19.9.0
html5lib==0.999
itsdangerous==0.24
jsonpatch==1.3
jsonpointer==1.0
kombu==4.2.1
Landscape-Client==14.12
Markdown==2.6.11
MarkupSafe==1.0
meld3==0.6.10
oauth==1.0.1
PAM==0.4.2
prettytable==0.7.2
pycurl==7.19.3
pyOpenSSL==0.13
pyserial==2.6
python-apt===0.9.3.5ubuntu3
python-debian===0.1.21-nmu2ubuntu2
pytz==2018.5
PyYAML==3.10
redis==2.10.6
requests==2.2.1
six==1.5.2
ssh-import-id==3.21
supervisor==3.0b2
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
urllib3==1.7.1
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.2.1
zope.interface==4.0.5
